app/backend-app/requirements.txt
# app/backend-app/requirements.txt

Flask==2.3.2
psycopg2-binary==2.9.9 # Use psycopg2-binary for simpler installation in Docker
gunicorn